highp test

+++ src/shaders.cpp
@@ -296,7 +296,7 @@ ColorShader::ColorShader() : AnyAminoShader() {
     //shaders
 //cbxx precision -> mediump
     fragmentShader = R"(
-        precision lowp float;
+        precision highp float;
 
         uniform vec4 color;
 
@@ -355,7 +355,7 @@ ColorLightingShader::ColorLightingShader() : ColorShader() {
     )";
 //cbxx precision
     fragmentShader = R"(
-        precision lowp float;
+        precision highp float;
 
         varying float lightFac;
         uniform vec4 color;
@@ -462,7 +462,7 @@ TextureShader::TextureShader() : AnyAminoShader() {
 //cbxx precision
     //supports opacity and discarding of fully transparent pixels
     fragmentShader = R"(
-        precision lowp float;
+        precision highp float;
 
         varying vec2 uv;
 
@@ -562,7 +562,7 @@ TextureClampToBorderShader::TextureClampToBorderShader() : TextureShader() {
     //Note: supports clamp to border, using transparent texture
 //cbxx precision
     fragmentShader = R"(
-        precision lowp float;
+        precision highp float;
 
         varying vec2 uv;
